+++ src/types/tree.d.ts
@@ -5,4 +5,21 @@ interface ITree {
   file: json;
 }
 
-export type { ITree };
+interface NodeData {
+  id: number;
+  pids?: string | string[] | number | number[];
+  gender: "male" | "female";
+  name: string;
+  photo?: string;
+  birthDate?: string;
+  deathDate?: string;
+  address?: string;
+  phone?: string;
+  email?: string;
+  occupation?: string;
+  note?: string;
+  mid?: number;
+  fid?: number;
+}
+
+export type { ITree, NodeData };
